+++ CHANGELOG.md
@@ -19,6 +19,12 @@ This driver uses semantic versioning:
 - A change in the major version (e.g. 1.Y.Z -> 2.0.0) indicates _breaking_
   changes that require changes in your code to upgrade.
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed default return type of AQL queries being `undefined` instead of `any` ([#797](https://github.com/arangodb/arangojs/issues/797))
+
 ## [8.4.0] - 2023-07-10
 
 ### Changed
@@ -1710,6 +1716,7 @@ For a detailed list of changes between pre-release versions of v7 see the
 
   Graph methods now only return the relevant part of the response body.
 
+[unreleased]: https://github.com/arangodb/arangojs/compare/v8.4.0...HEAD
 [8.4.0]: https://github.com/arangodb/arangojs/compare/v8.3.1...v8.4.0
 [8.3.1]: https://github.com/arangodb/arangojs/compare/v8.3.0...v8.3.1
 [8.3.0]: https://github.com/arangodb/arangojs/compare/v8.2.1...v8.3.0