+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This driver uses semantic versioning:
 
 - Implemented logging API (DE-144, DE-145, DE-146, DE-147)
 
+- Implemented async jobs management (DE-339)
+
 ## [8.4.1] - 2023-09-15
 
 ### Fixed